api-test-framework/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-test-framework"
version = "1.0.0"
description = "Production-grade, design-pattern-driven API Testing Framework"
requires-python = ">=3.11"
dependencies = [
    "pytest>=7.4,<9.0",
    "pytest-bdd>=7.0,<8.0",
    "pytest-xdist>=3.5,<4.0",
    "allure-pytest>=2.13,<3.0",
    "requests>=2.31,<3.0",
    "httpx>=0.27,<1.0",
    "pydantic[email]>=2.5,<3.0",
    "pydantic-settings>=2.1,<3.0",
    "jsonschema>=4.20,<5.0",
    "faker>=22.0,<30.0",
    "python-dotenv>=1.0,<2.0",
    "azure-eventhub>=5.11,<6.0",
    "confluent-kafka>=2.3,<3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2",
    "mypy>=1.8",
]

[tool.pytest.ini_options]
bdd_features_base_dir = "features"
addopts = [
    "--alluredir=reports/allure-results",
    "-p", "plugins.allure_plugin",
    "-p", "plugins.parallel_isolation",
    "-p", "plugins.request_logger",
    "-p", "plugins.api_version_plugin",
    "--strict-markers",
]
markers = [
    "smoke: Quick smoke tests for critical paths",
    "regression: Full regression test suite",
    "critical: Business-critical scenarios",
    "auth: Authentication and authorization tests",
    "users: User management tests",
    "events: Event-driven messaging tests",
    "api_version: Run against specific API version",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "scripts*", "reports*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true
